chore: moved team service
